requirements.txt
librosa
numpy
opencv-contrib-python
opencv-python
torch
torchvision
tqdm
numba
google-cloud-speech
google-cloud-storage
google-genai==1.12.1